+++ viadot/flows/azure_sql_transform.py
@@ -2,9 +2,7 @@ from typing import Any, Dict, List
 
 from prefect import Flow
 
-from ..tasks.azure_sql import AzureSQLDBQuery
-
-query_task = AzureSQLDBQuery()
+from viadot.tasks.azure_sql import AzureSQLDBQuery
 
 
 class AzureSQLTransform(Flow):
@@ -15,6 +13,7 @@ class AzureSQLTransform(Flow):
         sqldb_credentials_secret: str = None,
         vault_name: str = None,
         tags: List[str] = ["transform"],
+        timeout: int = 3600,
         *args: List[any],
         **kwargs: Dict[str, Any]
     ):
@@ -28,17 +27,20 @@ class AzureSQLTransform(Flow):
             with SQL db credentials (server, db_name, user, and password).
             vault_name (str, optional): The name of the vault from which to obtain the secret. Defaults to None.
             tags (list, optional): Tag for marking flow. Defaults to "transform".
+            timeout(int, optional): The amount of time (in seconds) to wait while running this task before
+                a timeout occurs. Defaults to 3600.
         """
         self.query = query
         self.tags = tags
         self.sqldb_credentials_secret = sqldb_credentials_secret
         self.vault_name = vault_name
-        self.tasks = [query_task]
+        self.timeout = timeout
 
         super().__init__(*args, name=name, **kwargs)
         self.gen_flow()
 
     def gen_flow(self) -> Flow:
+        query_task = AzureSQLDBQuery(timeout=self.timeout)
         query_task.bind(
             query=self.query,
             credentials_secret=self.sqldb_credentials_secret,