+++ viadot/flows/mindful_to_adls.py
@@ -33,6 +33,7 @@ class MindfulToADLS(Flow):
         region: Literal["us1", "us2", "us3", "ca1", "eu1", "au1"] = "eu1",
         file_extension: Literal["parquet", "csv"] = "csv",
         sep: str = "\t",
+        timeout: int = 3600,
         file_path: str = "",
         adls_file_path: str = None,
         adls_overwrite: bool = True,
@@ -53,6 +54,8 @@ class MindfulToADLS(Flow):
             region (Literal[us1, us2, us3, ca1, eu1, au1], optional): SD region from where to interact with the mindful API. Defaults to "eu1".
             file_extension (Literal[parquet, csv], optional): File extensions for storing responses. Defaults to "csv".
             sep (str, optional): Separator in csv file. Defaults to "\t".
+            timeout(int, optional): The amount of time (in seconds) to wait while running this task before
+                a timeout occurs. Defaults to 3600.
             file_path (str, optional): Path where to save the file locally. Defaults to ''.
             adls_file_path (str, optional): The destination path at ADLS. Defaults to None.
             adls_overwrite (bool, optional): Whether to overwrite files in the data lake. Defaults to True.
@@ -69,6 +72,7 @@ class MindfulToADLS(Flow):
         self.file_extension = file_extension
         self.sep = sep
         self.file_path = file_path
+        self.timeout = timeout
 
         self.adls_file_path = adls_file_path
         self.adls_overwrite = adls_overwrite
@@ -79,7 +83,7 @@ class MindfulToADLS(Flow):
         self.mind_flow()
 
     def mind_flow(self) -> Flow:
-        to_csv = MindfulToCSV()
+        to_csv = MindfulToCSV(timeout=self.timeout)
 
         file_names = to_csv.bind(
             credentials_mindful=self.credentials_mindful,
@@ -101,6 +105,7 @@ class MindfulToADLS(Flow):
             adls_file_path=self.adls_file_path,
             adls_sp_credentials_secret=self.adls_sp_credentials_secret,
             adls_overwrite=self.adls_overwrite,
+            task_timeout=self.timeout,
             flow=self,
         )